refactor(libs): extract runtime-agnostic Prisma client to @libs/prisma (#1090)
Foundation for the worker/API decoupling (step 1 of the dual-validated plan).
Splits PrismaService so workers stop importing @api source just to obtain a
Prisma client.

The problem: workers bound SERVER_TOKENS.prisma to the @api PrismaService, so
every worker wiring module imported @api — the boundary guard only blocks
@server→@api, never workers→@api, so #1090's zero-worker-@api goal was
unreachable while the only concrete client lived in @api.

The split:
- packages/libs/prisma/prisma-pg-config.ts — pure connection-string parsing,
  node-postgres pool sizing, and RDS TLS/CA resolution (verbatim; zero @api).
- packages/libs/prisma/prisma.service.ts — runtime-agnostic base PrismaService
  (connection lifecycle + optional query-event listener via constructor option).
- packages/libs/prisma/prisma.module.ts — @Global module for non-API runtimes.
- apps/server/api/.../prisma.service.ts — now an ApiPrismaService that extends
  the @libs base and wires the request-performance instrumentation via
  onQueryEvent. Keeps the `PrismaService` export, so all ~290 API importers and
  the @api PrismaModule are untouched.

Wiring:
- Workers app.module + 16 worker files re-point @api prisma → @libs/prisma
  (the base client; the API-only AsyncLocalStorage instrumentation is a verified
  no-op in workers anyway — no request context).
- Move the createPrismaPgConfig tests to @libs alongside the factory.
- Prune the 2 prisma entries from the workers @api-import baseline (178→176).

Verified: workers import zero @api prisma; @libs/prisma imports zero @api; only
workers load the @libs PrismaModule (API keeps its instrumented adapter under
the same token, so no second connection pool); baseline set == worker usage.
