+++ apps/server/workers/src/crons/agent/cron.proactive-agent.service.spec.ts
@@ -4,9 +4,9 @@ import { CreditsUtilsService } from '@api/collections/credits/services/credits.u
 import { OrganizationSettingsService } from '@api/collections/organization-settings/services/organization-settings.service';
 import { AgentRunQueueService } from '@api/queues/agent-run/agent-run-queue.service';
 import { CacheService } from '@api/services/cache/services/cache.service';
-import { PrismaService } from '@api/shared/modules/prisma/prisma.service';
 import { AgentRunFrequency, AgentType } from '@genfeedai/enums';
 import { LoggerService } from '@libs/logger/logger.service';
+import { PrismaService } from '@libs/prisma/prisma.service';
 import { Test, TestingModule } from '@nestjs/testing';
 import { CronProactiveAgentService } from '@workers/crons/agent/cron.proactive-agent.service';
