+++ apps/server/workers/src/crons/model-deprecation/cron.model-deprecation.service.spec.ts
@@ -1,7 +1,7 @@
 import { ModelsService } from '@api/collections/models/services/models.service';
-import { PrismaService } from '@api/shared/modules/prisma/prisma.service';
 import { ModelCategory, WorkflowStatus } from '@genfeedai/enums';
 import { LoggerService } from '@libs/logger/logger.service';
+import { PrismaService } from '@libs/prisma/prisma.service';
 import { Test, TestingModule } from '@nestjs/testing';
 import { CronModelDeprecationService } from '@workers/crons/model-deprecation/cron.model-deprecation.service';
