+++ apps/server/workers/src/crons/workflows/cron.workflows.service.ts
@@ -1,7 +1,6 @@
 import { CreditsUtilsService } from '@api/collections/credits/services/credits.utils.service';
 import { CacheService } from '@api/services/cache/services/cache.service';
 import { NotificationsService } from '@api/services/notifications/notifications.service';
-import { PrismaService } from '@api/shared/modules/prisma/prisma.service';
 import {
   WorkflowRecurrenceType,
   WorkflowStatus,
@@ -10,6 +9,7 @@ import {
 } from '@genfeedai/enums';
 import type { Workflow } from '@genfeedai/prisma';
 import { LoggerService } from '@libs/logger/logger.service';
+import { PrismaService } from '@libs/prisma/prisma.service';
 import { Injectable } from '@nestjs/common';
 import { Cron, CronExpression } from '@nestjs/schedule';
 import { GenerateArticleTask } from '@workers/crons/workflows/task-types/generate-article.task';
